requirements.txt
transformers==4.49.0
torch==2.6.0
torchvision==0.21.0
torchaudio==2.6.0
pillow==11.1.0
numpy==2.0.2
matplotlib==3.9.4
fastapi==0.115.11
uvicorn==0.34.0
python-multipart==0.0.20
cloudinary==1.42.2
python-dotenv==1.0.1
einops==0.8.1
timm==1.0.15
pydantic==2.10.6
annotated-types==0.7.0
anyio==4.8.0
triton==3.2.0